Testing if a class component using IntersectionObserver in the lifecycle method will successfully build in Netlify.
