+++ src/components/index.js
@@ -1,3 +1,3 @@
 export { default as Head } from './Head'
 export { default as Navbar } from './Navbar'
-export { default as ProjectCard } from './ProjectCard'
+export { default as ProjectList } from './ProjectList'
