+++ src/views/Projects.js
@@ -1,7 +1,6 @@
 // React modules
-import React, { useState, useEffect, createRef } from 'react'
-import { ProjectCard } from '@components'
-import { useStaticQuery, graphql } from 'gatsby'
+import React, { Component } from 'react'
+import { ProjectList } from '@components'
 
 // Styling modules
 import styled, { css } from 'styled-components'
@@ -9,7 +8,6 @@ import { theme, Section } from '@theme'
 
 // Misc. modules
 import content from '@content/projects.yml'
-import { generateKey } from '@utils'
 
 // =============================================================================
 
@@ -20,93 +18,56 @@ import { generateKey } from '@utils'
  * @example
  * imgSrc: project-mockup-small.png
  */
-const Projects = () => {
-    /**
-     * Possible values:
-     * - srcWebp
-     * - src
-     */
-    const endAttr = 'srcWebp'
+class Projects extends Component {
+    observer = null
+    threshold = [0.7, 0.5, 0.5, 0.5]
+    componentDidMount() {
+        const { handleObservedEntries, threshold } = this
+        this.observer = new IntersectionObserver(handleObservedEntries, {
+            threshold,
+        })
+    }
 
-    const data = useStaticQuery(
-        graphql`
-            query loadProjectMockups {
-                allImageSharp {
-                    edges {
-                        node {
-                            fluid(maxWidth: 696, quality: 100) {
-                                originalName
-                                srcWebp
-                            }
-                        }
-                    }
-                }
-            }
-        `
-    )
+    componentWillUnmount() {
+        this.observer = null
+    }
 
-    /**
-     * Creates a hash map where each key is the image file name and the value
-     * is an object containing the `src` and `originalName` properties of
-     * the image.
-     */
-    const imageHashMap = data.allImageSharp.edges.reduce((hashMap, edge) => {
-        const imageDetails = edge.node.fluid
-        if (hashMap[imageDetails.originalName]) {
-            // Do something with duplicate entries
-        } else {
-            hashMap[imageDetails.originalName] = { ...imageDetails }
-        }
-        return hashMap
-    }, {})
+    handleObservedEntries = (entries) => {
+        console.log(entries)
+    }
 
-    /**
-     * Updates the `imgSrc` property of each project object literal to the
-     * `src` value generated by Gatsby.
-     */
-    const projects = content.projects.map((project) => ({
-        ...project,
-        imgSrc: imageHashMap[project.imgSrc]
-            ? imageHashMap[project.imgSrc][endAttr]
-            : imageHashMap['project-mockup-sample.png'][endAttr],
-    }))
+    observeElement = () => {}
 
-    /**
-     * This portion handles the opacity of each project card. It uses an
-     * intersection observer, observing each card.
-     */
+    render() {
+        /**
+         * This portion handles the opacity of each project card. It uses an
+         * intersection observer, observing each card.
+         */
 
-    const observer = new IntersectionObserver(
-        (entries) => {
-            entries.forEach((entry) => {
-                if (entry.isIntersecting) {
-                    entry.target.classList.add('opaque')
-                } else {
-                    entry.target.classList.remove('opaque')
-                }
-            })
-        },
-        { threshold: [0.7, 0.5, 0.5, 0.5] }
-    )
+        // useEffect(() => {
+        //     const observer = new IntersectionObserver(
+        //         (entries) => {
+        //             entries.forEach((entry) => {
+        //                 if (entry.isIntersecting) {
+        //                     entry.target.classList.add('opaque')
+        //                 } else {
+        //                     entry.target.classList.remove('opaque')
+        //                 }
+        //             })
+        //         },
+        //         { threshold: [0.7, 0.5, 0.5, 0.5] }
+        //     )
+        // }, [])
 
-    const observeElement = (ref) => observer.observe(ref)
+        const observeElement = () => console.log('hi')
 
-    return (
-        <Section id="projects" additionalStyles={additionalSectionStyles}>
-            <Heading>{content.heading}</Heading>
-            <ProjectList>
-                {projects.map((project, i) => {
-                    return (
-                        <ProjectCard
-                            key={generateKey(project.name, i)}
-                            observeElement={observeElement}
-                            {...project}
-                        />
-                    )
-                })}
-            </ProjectList>
-        </Section>
-    )
+        return (
+            <Section id="projects" additionalStyles={additionalSectionStyles}>
+                <Heading>{content.heading}</Heading>
+                <ProjectList observeElement={observeElement} />
+            </Section>
+        )
+    }
 }
 
 // =============================================================================
@@ -188,24 +149,6 @@ const Heading = styled('h2')`
             left: unset;
             height: 6px;
         }
-    }
-`
-
-// =============================================================================
-
-const ProjectList = styled('ul')`
-    margin-top: 60px;
-
-    @media only screen and (min-width: 768px) {
-        margin-top: 90px;
-    }
-
-    @media only screen and (min-width: 1024px) {
-        margin-top: 100px;
-    }
-
-    @media only screen and (min-width: 1440px) {
-        margin-top: 130px;
     }
 `
